Fixed bug with rand of axis values.

+++ test_robot_module.cpp
@@ -206,7 +206,7 @@ void TestRobot::axisControl(system_value axis_index, variable_value value) {
 		case 3: { name = "Z"; break; }
 		default: { name = "O_o"; break; };
 	}
-	parent->colorPrintf(ConsoleColor(ConsoleColor::green), "change axis value: %s = %d\n", name, (int) value);
+	parent->colorPrintf(ConsoleColor(ConsoleColor::green), "change axis value: %s = %f\n", name, value);
 }
 
 PREFIX_FUNC_DLL RobotModule* getRobotModuleObject() {
